Move some classes.
