+++ agents/gcs-to-gcs/src/main/scala/org/broadinstitute/transporter/api/GcsToGcsRequest.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.transporter.transfer
+package org.broadinstitute.transporter.api
 
 import io.circe.Decoder
 import io.circe.derivation.deriveDecoder
